[김동욱] sprint8

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -2,18 +2,24 @@ package com.sprint.mission.discodeit.service.basic;
 
 import com.sprint.mission.discodeit.dto.message.request.MessageCreateRequest;
 import com.sprint.mission.discodeit.dto.message.request.MessageUpdateRequest;
-import com.sprint.mission.discodeit.dto.message.response.AdvancedJpaPageResponse;
-import com.sprint.mission.discodeit.dto.message.response.JpaMessageResponse;
-import com.sprint.mission.discodeit.dto.message.response.JpaPageResponse;
-import com.sprint.mission.discodeit.entity.*;
-import com.sprint.mission.discodeit.mapper.advanced.AdvancedMessageMapper;
-import com.sprint.mission.discodeit.mapper.original.PageResponseMapper;
-import com.sprint.mission.discodeit.repository.jpa.*;
+import com.sprint.mission.discodeit.dto.message.response.PageResponse;
+import com.sprint.mission.discodeit.dto.message.response.MessageResponse;
+import com.sprint.mission.discodeit.entity.BinaryContent;
+import com.sprint.mission.discodeit.entity.Channel;
+import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.exception.channelException.ChannelNotFoundException;
+import com.sprint.mission.discodeit.exception.messageException.MessageNotFoundException;
+import com.sprint.mission.discodeit.exception.userException.UserNotFoundException;
+import com.sprint.mission.discodeit.mapper.MessageMapper;
+import com.sprint.mission.discodeit.repository.jpa.JpaBinaryContentRepository;
+import com.sprint.mission.discodeit.repository.jpa.JpaChannelRepository;
+import com.sprint.mission.discodeit.repository.jpa.JpaMessageRepository;
+import com.sprint.mission.discodeit.repository.jpa.JpaUserRepository;
 import com.sprint.mission.discodeit.service.MessageService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +27,10 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.time.Instant;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 /**
  * packageName    : com.sprint.mission.discodeit.service.basic
@@ -43,65 +52,70 @@ public class BasicMessageService implements MessageService {
     private final JpaChannelRepository channelRepository;
     private final JpaUserRepository userRepository;
     private final JpaBinaryContentRepository binaryContentRepository;
-//    private final JpaReadStatusRepository readStatusRepository;
-//    private final MessageMapper messageMapper;
-    private final AdvancedMessageMapper messageMapper;
+    private final MessageMapper messageMapper;
     private final BinaryContentStorage binaryContentStorage;
-//    private final PageResponseMapper pageResponseMapper;
-
 
     @Override
-    public AdvancedJpaPageResponse findAllByChannelIdAndCursor(UUID channelId, Instant cursor, Pageable pageable) {
+    public PageResponse findAllByChannelIdAndCursor(UUID channelId, Instant cursor, Pageable pageable) {
 
-        List<Message> messages;
-        if(cursor == null) {
-//            messages = messageRepository.findAllByChannelIdOrderByCreatedAt(channelId, pageable);
-            messages = messageRepository.findAllByChannelId(channelId, pageable);
-        }else {
-            messages = messageRepository.findByChannelIdAndCreatedAtBeforeOrderByCreatedAtDesc(channelId, cursor, pageable);
-        }
+        List<Message> messages = messageRepository
+            .findSliceByCursor(channelId, cursor, pageable);
+
         boolean hasNext = messages.size() > pageable.getPageSize();
-        Long totalElements = messageRepository.countByChannelId(channelId);
-        Instant nextCursor = null;
-        if(hasNext) {
+        if (hasNext) {
             messages = messages.subList(0, pageable.getPageSize());
-            nextCursor = messages.get(messages.size() - 1).getCreatedAt();
         }
+        Instant nextCursor = hasNext ? messages.get(messages.size() - 1).getCreatedAt() : null;
 
-        List<JpaMessageResponse> messagesDto = messages.stream().map(message -> messageMapper.toDto(message)).toList();
+        List<MessageResponse> dtoList = messages.stream()
+            .map(messageMapper::toDto)
+            .toList();
 
-        AdvancedJpaPageResponse response = AdvancedJpaPageResponse.builder()
-                .content(messagesDto)
-                .nextCursor(nextCursor)
-                .size(messages.size())
-                .hasNext(hasNext)
-                .totalElements(totalElements)
-                .build();
-        return response;
+        return PageResponse.builder()
+            .content(dtoList)
+            .nextCursor(nextCursor)
+            .size(dtoList.size())
+            .hasNext(hasNext)
+            .totalElements(messageRepository.countByChannelId(channelId))
+            .build();
     }
 
-    // for(BinaryContent 생성 -> 이미지 저장 -> BinaryContent Id 리스트로 저장)  -> 메세지 생성
-    // public 방일경우 작성을 해도 readstatus가 없음 최소 1회는 등록을 해야 하고 유저가 방에 있는지 확인 가능한 로직이 필요함
-    // binary content
+    /**
+     * for(BinaryContent 생성 -> 이미지 저장 -> BinaryContent Id 리스트로 저장)  -> 메세지 생성
+     * public 방일경우 작성을 해도 readstatus가 없음 최소 1회는 등록을 해야 하고 유저가 방에 있는지 확인 가능한 로직이 필요함
+     * binary content
+     * @param request
+     * @param fileList
+     * @return
+     */
     @Override
-    public JpaMessageResponse createMessage(MessageCreateRequest request, List<MultipartFile> fileList) {
-
-        Channel channel = channelRepository.findById(request.channelId()).orElseThrow(() -> new NoSuchElementException("channel with id " + request.channelId() + " not found"));
-        User user = userRepository.findById(request.authorId()).orElseThrow(() -> new NoSuchElementException("author with id " + request.authorId() + " not found"));
+    public MessageResponse createMessage(MessageCreateRequest request, List<MultipartFile> fileList) {
+        Channel channel = channelRepository.findById(request.channelId()).orElseThrow(() -> new ChannelNotFoundException(Map.of("channelId",request.channelId())));
+        User user = userRepository.findById(request.authorId()).orElseThrow(() -> new UserNotFoundException(Map.of("userId",request.authorId())));
 
         // for(BinaryContent 생성 -> 이미지 저장 -> BinaryContent Id 리스트로 저장)
         List<BinaryContent> attachments = new ArrayList<>();
         if (hasValue(fileList)) {
             for (MultipartFile file : fileList) {
+
                 // BinaryContent 생성
                 BinaryContent attachment;
                 try {
+                    String originalFilename = file.getOriginalFilename();
+                    String extension = "";
+
+                    int dotIndex = originalFilename.lastIndexOf(".");
+                    if (dotIndex != -1 && dotIndex < originalFilename.length() - 1) {
+                        extension = originalFilename.substring(dotIndex);
+                    } else {
+                        extension = "";
+                    }
                     BinaryContent binaryContent = BinaryContent.builder()
-                            .fileName(file.getOriginalFilename())
-                            .size((long) file.getBytes().length)
-                            .contentType(file.getContentType())
-                            .extension(file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf(".")))
-                            .build();
+                        .fileName(originalFilename)
+                        .size((long) file.getBytes().length)
+                        .contentType(file.getContentType())
+                        .extension(extension)
+                        .build();
                     attachment = binaryContentRepository.save(binaryContent);
                     binaryContentStorage.put(binaryContent.getId(),file.getBytes());
 
@@ -110,11 +124,6 @@ public class BasicMessageService implements MessageService {
                 }
 
                 attachments.add(attachment);
-                try {
-                    binaryContentStorage.put(attachment.getId(), file.getBytes());
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
             }
         }
 
@@ -127,25 +136,26 @@ public class BasicMessageService implements MessageService {
                 .build();
         messageRepository.save(message);
 
-        JpaMessageResponse response = messageMapper.toDto(message);
+        MessageResponse response = messageMapper.toDto(message);
 
         return response;
         // for(BinaryContent 생성 -> 이미지 저장 -> BinaryContent Id 리스트로 저장)  -> 메세지 생성
     }
 
     @Override
-    public boolean deleteMessage(UUID messageId) {
-        Optional.ofNullable(messageId).orElseThrow(() -> new IllegalArgumentException("require message Id : BasicMessageService.deleteMessage"));
+    public void deleteMessage(UUID messageId) {
+        if(!messageRepository.existsById(messageId)) {
+            throw new MessageNotFoundException(Map.of("messageId",messageId));
+        }
         messageRepository.deleteById(messageId);
-        return true;
     }
 
     @Override
-    public JpaMessageResponse updateMessage(UUID messageId, MessageUpdateRequest request) {
-        Message message = messageRepository.findById(messageId).orElseThrow(() -> new NoSuchElementException("message with id " + messageId + " not found"));
+    public MessageResponse updateMessage(UUID messageId, MessageUpdateRequest request) {
+        Message message = messageRepository.findById(messageId).orElseThrow(() -> new MessageNotFoundException(Map.of("messageId", messageId)));
         message.setContent(request.newContent());
 
-        JpaMessageResponse response = messageMapper.toDto(message);
+        MessageResponse response = messageMapper.toDto(message);
         return response;
     }
